merge

+++ evaluate.py
@@ -114,7 +114,7 @@ def eval(model, model_dir, weight_file_path, store_imgs, dataset_path, ex_id, th
             # I.paste(Image.blend(img.convert("L"), label_img.convert("L"), 0.2), (img.size[0]*3, 0))
             # I.paste(Image.blend(img.convert("L"), pred_img.convert("L"), 0.2), (img.size[0]*4, 0))
             
-            I = cv.cvtColor(img, cv.COLOR_GRAY2RGB)
+            I = cv.cvtColor(np.float32(img), cv.COLOR_GRAY2RGB)
 
             I[label_img == 255, :, :] =  255
             I[:, pred_img == 255, :] =  255
